In this problem you need to implement a function that shifts the contents of an array to the left by a given number of positions (cyclic shift).
The function takes as input an array, its size, and the value of the shift. For example, if int a[] = { 1, 2, 3, 4, 5 }; and it is necessary to shift it cyclically to the left by 2 positions, the output will be numbers in this order: 3, 4, 5, 1, 2.

Note that the shift value may be zero, or it may be larger than the array size, all these cases must be taken into account.