Implement a getline function that reads the input stream character by character until it reaches the end of the stream or meets a line break character ('\n'), and returns a C-style string with the read characters.
Note that since the input size is not known in advance, you will need to reallocate memory during the reading process if there are more characters in the input stream than you expect.

The memory returned from the function will be released by the delete[] operator. The line break character ('\n') does not need to be added to the string, but don't forget that there must be a null character at the end of the C-style string.