In this assignment, you are asked to develop a hierarchy of Expression classes to represent arithmetic expressions. Specifically, you need to define three classes: Expression - the base class of the hierarchy, Number - to represent numbers, and BinaryOperation - a class describing a binary operation (+, -, * or /).
The Number class has to store a value of the double type.

The BinaryOperation class should store a pointer to the left and right operands, which are themselves arithmetic expressions, and the type of operation (+, -, *, or /) to be performed on them.

All classes should have a method evaluate that returns a value of type double - the value of the corresponding arithmetic expression, for example, the value of an instance of type Number is the number it stores, and if you have a BinaryOperation object with a + operation, you need to calculate the values of the left and right operands and return their sum.

In this assignment, you need to put the virtual keyword where you need it, define the evaluate method where it's missing, and implement destructors where you need them.

When doing this task, keep in mind that when you destroy a BinaryOperation object, it is responsible for destroying the left and right operands (ensuring that they are allocated in dynamic memory).