You are given a Rational class that describes a rational number. It has add, sub, mul, and div methods that add to it, subtract from it, multiply by it, and divide by it, respectively. It also has a method neg, which reverses the sign.
You need to define +=, -=, *=, /= operators for the Rational class so that they can take both objects of type Rational and integers as arguments. You also need to define unary minus and plus operators for the Rational class.

Define addition, subtraction, multiplication and division operators for the Rational class so that objects of Rational type can be added (subtracted, multiplied and divided) not only with each other but also with integers.

Implement operators <, <=, >, >=, ==, != for Rational class so that it is possible to compare Rational class objects not only with each other but also with integers.

Add a cast to double operator to the Rational class.