Implement a template version of the Array class.
Implement a template function minimum that finds the minimum element that is stored in an instance of the Array template class, and the Array template parameter can be arbitrary.

Write a template function (or several) flatten that takes as input such a "multidimensional" Array of unknown nesting depth and outputs to the out stream all elements stored at the lowest level through a space.